chore: regen artifacts
