+++ tests_generated/test_all_default_parameters_off_jetbrains/backend/pyproject.toml
@@ -12,6 +12,7 @@ algokit-utils = "^2.2.0"
 python-dotenv = "^1.0.0"
 
 [tool.poetry.group.dev.dependencies]
+setuptools = "^69.0.2" # Adding explicitly to work around pyteal https://github.com/algorand/pyteal/issues/712
 
 [build-system]
 requires = ["poetry-core"]
