+++ tests_generated/test_all_default_parameters_off_jetbrains/frontend/.copier-answers.yml
@@ -1,5 +1,5 @@
 # Changes here will be overwritten by Copier; NEVER EDIT MANUALLY
-_commit: 0.9.1
+_commit: 0.9.3
 _src_path: gh:algorandfoundation/algokit-react-frontend-template
 algod_port: 4001
 algod_server: http://localhost
