+++ tests_generated/test_all_default_parameters_off_jetbrains/frontend/package.json
@@ -6,31 +6,32 @@
     "email": "None"
   },
   "private": true,
+  "type": "module",
   "engines": {
     "node": ">=18.0"
   },
   "devDependencies": {
-    "@types/node": "18.17.14",
-    "@types/react": "18.2.11",
-    "@types/react-dom": "18.2.4",
-    "@vitejs/plugin-react-swc": "3.3.2",
-    "autoprefixer": "10.4.14",
-    "ts-node": "10.9.1",
-    "typescript": "5.1.6",
-    "vite": "4.4.9"
+    "@types/node": "^18.17.14",
+    "@types/react": "^18.2.11",
+    "@types/react-dom": "^18.2.4",
+    "@vitejs/plugin-react": "^4.2.1",
+    "autoprefixer": "^10.4.14",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.1.6",
+    "vite": "^5.0.0"
   },
   "dependencies": {
     "@walletconnect/modal-sign-html": "^2.6.1",
-    "@algorandfoundation/algokit-utils": "^4.1.0",
+    "@algorandfoundation/algokit-utils": "^5.0.0",
     "@blockshake/defly-connect": "^1.1.6",
     "@daffiwallet/connect": "^1.0.3",
     "@perawallet/connect": "^1.3.1",
-    "@txnlab/use-wallet": "^2.1.1",
-    "algosdk": "^2.5.0",
-    "daisyui": "^3.1.0",
+    "@txnlab/use-wallet": "^2.4.0",
+    "algosdk": "^2.7.0",
+    "daisyui": "^4.0.0",
     "notistack": "^3.0.1",
     "react": "^18.2.0",
-    "react-dom": "18.2.0",
+    "react-dom": "^18.2.0",
     "tslib": "^2.6.2"
   },
   "scripts": {
