+++ tests_generated/test_all_default_parameters_off_jetbrains/frontend/vite.config.ts
@@ -1,4 +1,4 @@
-import react from '@vitejs/plugin-react-swc'
+import react from '@vitejs/plugin-react'
 import { defineConfig } from 'vite'
 
 // https://vitejs.dev/config/
