+++ tests_generated/test_all_default_parameters_on_production/.github/workflows/frontend-cd.yaml
@@ -56,4 +56,5 @@ jobs:
         env:
           NETLIFY_SITE_ID: ${{ secrets.NETLIFY_SITE_ID }}
           NETLIFY_AUTH_TOKEN: ${{ secrets.NETLIFY_AUTH_TOKEN }}
+          # Set your vercel project env variables on your github repository (see README in frontend folder for more info)
