+++ tests_generated/test_all_default_parameters_on_production/backend/.tours/getting-started-with-your-algokit-project.tour
@@ -15,7 +15,7 @@
     {
       "file": "smart_contracts/__main__.py",
       "description": "This is the main entry point for building your smart contracts. The default template includes a starter 'Hello World' contract that is deployed via the `algokit-utils` package (either `ts` or `py`, depending on your choice). To create a new smart contract, you can use the [`algokit generate`](https://github.com/algorandfoundation/algokit-cli/blob/main/docs/features/generate.md) command and invoke a pre-bundled generator template by running `algokit generate smart-contract`. This action will create a new folder in the `smart_contracts` directory, named after your project. Each folder contains a `contract.py` file, which is the entry point for your contract implementation, and `deploy_config.py` | `deployConfig.ts` files (depending on the language chosen for the template), that perform the deployment of the contract.",
-      "line": 25
+      "line": 26
     },
     {
       "file": "smart_contracts/hello_world/deploy_config.py",
