+++ tests_generated/test_all_default_parameters_on_production/backend/smart_contracts/__main__.py
@@ -8,8 +8,9 @@ from smart_contracts.config import contracts
 from smart_contracts.helpers.build import build
 from smart_contracts.helpers.deploy import deploy
 
-# # Uncomment the following lines to enable auto generation of AVM Debugger compliant sourceMap and simulation trace file.
-# # Learn more about using AlgoKit AVM Debugger to debug your TEAL source codes and inspect various kinds of Algorand transactions in atomic groups -> https://github.com/algorandfoundation/algokit-avm-vscode-debugger
+# Uncomment the following lines to enable auto generation of AVM Debugger compliant sourcemap and simulation trace file.
+# Learn more about using AlgoKit AVM Debugger to debug your TEAL source codes and inspect various kinds of
+# Algorand transactions in atomic groups -> https://github.com/algorandfoundation/algokit-avm-vscode-debugger
 # from algokit_utils.config import config
 # config.configure(debug=True, trace_all=True)
 logging.basicConfig(
