+++ tests_generated/test_all_default_parameters_on_production/backend/tests/hello-world.spec.ts
@@ -29,6 +29,7 @@ describe('hello world contract', () => {
       onSchemaBreak: 'replace',
       onUpdate: 'update',
     })
+    
     return { client }
   }
