+++ tests_generated/test_all_default_parameters_on_production/frontend/.prettierrc.cjs
