+++ tests_generated/test_all_default_parameters_on_production/frontend/package.json
@@ -6,42 +6,43 @@
     "email": "None"
   },
   "private": true,
+  "type": "module",
   "engines": {
     "node": ">=18.0"
   },
   "devDependencies": {
-    "@types/node": "18.17.14",
-    "@types/react": "18.2.11",
-    "@types/react-dom": "18.2.4",
-    "@vitejs/plugin-react-swc": "3.3.2",
-    "autoprefixer": "10.4.14",
-    "eslint": "8.42.0",
-    "eslint-config-prettier": "8.8.0",
-    "eslint-plugin-prettier": "5.0.0",
-    "@typescript-eslint/eslint-plugin": "6.5.0",
-    "@typescript-eslint/parser": "6.5.0",
+    "@types/node": "^18.17.14",
+    "@types/react": "^18.2.11",
+    "@types/react-dom": "^18.2.4",
+    "@vitejs/plugin-react": "^4.2.1",
+    "autoprefixer": "^10.4.14",
+    "eslint": "^8.42.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-prettier": "^5.0.0",
+    "@typescript-eslint/eslint-plugin": "^6.5.0",
+    "@typescript-eslint/parser": "^6.5.0",
     "postcss": "^8.4.24",
     "tailwindcss": "3.3.2",
     "ts-jest": "^29.1.1",
     "@types/jest": "29.5.2",
-    "ts-node": "10.9.1",
-    "typescript": "5.1.6",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.1.6",
     "@playwright/test": "^1.35.0",
     "playwright": "^1.35.0",
-    "vite": "4.4.9"
+    "vite": "^5.0.0"
   },
   "dependencies": {
     "@walletconnect/modal-sign-html": "^2.6.1",
-    "@algorandfoundation/algokit-utils": "^4.1.0",
+    "@algorandfoundation/algokit-utils": "^5.0.0",
     "@blockshake/defly-connect": "^1.1.6",
     "@daffiwallet/connect": "^1.0.3",
     "@perawallet/connect": "^1.3.1",
-    "@txnlab/use-wallet": "^2.1.1",
-    "algosdk": "^2.5.0",
-    "daisyui": "^3.1.0",
+    "@txnlab/use-wallet": "^2.4.0",
+    "algosdk": "^2.7.0",
+    "daisyui": "^4.0.0",
     "notistack": "^3.0.1",
     "react": "^18.2.0",
-    "react-dom": "18.2.0",
+    "react-dom": "^18.2.0",
     "tslib": "^2.6.2"
   },
   "scripts": {
